+++ app/components/UI/Predict/components/PredictGameDetailsContent/PredictGameOutcomesTab.tsx
@@ -1,600 +1,16 @@
-import React, {
-  memo,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import React, { memo } from 'react';
 import { Box } from '@metamask/design-system-react-native';
 import type {
-  GetPriceResponse,
   PredictMarketGame,
   PredictOutcome,
   PredictOutcomeGroup,
   PredictOutcomeToken,
-  PriceQuery,
 } from '../../types';
-import { useLiveMarketPrices } from '../../hooks/useLiveMarketPrices';
-import { usePredictPrices } from '../../hooks/usePredictPrices';
-import type { PredictBetButtonVariant } from '../PredictActionButtons/PredictActionButtons.types';
-import PredictSportOutcomeCard, {
-  type PredictSportOutcomeButton,
-} from '../PredictSportOutcomeCard';
-import { formatVolume } from '../../utils/format';
-import { getPredictBuyPrice, isValidPrice } from '../../utils/prices';
-import { isMoneylineLikeMarketType } from '../../constants/sports';
-import { usePredictPreviewSheet } from '../../contexts';
-import { strings } from '../../../../../../locales/i18n';
-import Logger from '../../../../../util/Logger';
 import { PREDICT_GAME_DETAILS_CONTENT_TEST_IDS } from './PredictGameDetailsContent.testIds';
+import { OutcomesContent } from './PredictGameOutcomeCards';
+import { usePricedOutcomeGroup } from './usePricedOutcomeGroup';
 
-const I18N_PREFIX = 'predict.sports_market_types';
-const MISSING_TRANSLATION_PREFIX = '[missing';
-const loggedMissingTranslationKeys = new Set<string>();
-const PRICE_POLL_INTERVAL_MS = 2000;
-const EMPTY_PRICE_QUERIES: PriceQuery[] = [];
-
-const flattenGroupOutcomes = (group: PredictOutcomeGroup): PredictOutcome[] => [
-  ...group.outcomes,
-  ...(group.subgroups?.flatMap((subgroup) => subgroup.outcomes) ?? []),
-];
-
-const buildPriceQueries = (outcomes: PredictOutcome[]): PriceQuery[] =>
-  outcomes.flatMap((outcome) =>
-    outcome.tokens.map((token) => ({
-      marketId: outcome.marketId,
-      outcomeId: outcome.id,
-      outcomeTokenId: token.id,
-    })),
-  );
-
-const getOutcomeBuyPrice = (
-  token: PredictOutcomeToken,
-  prices: GetPriceResponse,
-): number => {
-  const restEntry = prices.results.find(
-    (result) => result.outcomeTokenId === token.id,
-  );
-  if (isValidPrice(restEntry?.entry.sell)) {
-    return restEntry.entry.sell;
-  }
-
-  return getPredictBuyPrice(token, undefined, prices) ?? token.price;
-};
-
-const applyPricesToOutcomes = (
-  outcomes: PredictOutcome[],
-  prices: GetPriceResponse,
-  previousById?: Map<string, PredictOutcome>,
-): PredictOutcome[] =>
-  outcomes.map((outcome) => {
-    const previousOutcome = previousById?.get(outcome.id);
-    let changed = !previousOutcome;
-
-    const tokens = outcome.tokens.map((token) => {
-      const price = getOutcomeBuyPrice(token, prices);
-      const previousToken = previousOutcome?.tokens.find(
-        (t) => t.id === token.id,
-      );
-      if (previousToken && previousToken.price === price) {
-        return previousToken;
-      }
-      changed = true;
-      return { ...token, price };
-    });
-
-    if (previousOutcome && !changed) {
-      return previousOutcome;
-    }
-    return { ...outcome, tokens };
-  });
-
-const applyPricesToGroup = (
-  group: PredictOutcomeGroup,
-  prices: GetPriceResponse,
-  previousGroup: PredictOutcomeGroup | null,
-  previousPricedOutcomes: PredictOutcome[],
-): PredictOutcomeGroup => {
-  const canReusePreviousPrices = previousGroup === group;
-  const previousById = canReusePreviousPrices
-    ? new Map(previousPricedOutcomes.map((outcome) => [outcome.id, outcome]))
-    : undefined;
-
-  const pricedOutcomes = applyPricesToOutcomes(
-    group.outcomes,
-    prices,
-    previousById,
-  );
-  const pricedSubgroups = group.subgroups?.map((subgroup) => ({
-    ...subgroup,
-    outcomes: applyPricesToOutcomes(subgroup.outcomes, prices, previousById),
-  }));
-
-  return {
-    ...group,
-    outcomes: pricedOutcomes,
-    ...(pricedSubgroups && { subgroups: pricedSubgroups }),
-  };
-};
-
-const toTitleCase = (str: string): string =>
-  str
-    .split('_')
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(' ');
-
-const isMissingTranslation = (value: string, key: string): boolean =>
-  value === key || value.startsWith(MISSING_TRANSLATION_PREFIX);
-
-const logMissingSportsMarketTypeTranslation = (
-  key: string,
-  type: string,
-): void => {
-  if (loggedMissingTranslationKeys.has(key)) return;
-
-  loggedMissingTranslationKeys.add(key);
-  const message = `Missing Predict sports market type translation: ${key}`;
-  Logger.error(new Error(message), {
-    message,
-    context: { key, type },
-  });
-};
-
-const getTranslatedSportsMarketTypeLabel = (
-  type: string,
-): string | undefined => {
-  const key = `${I18N_PREFIX}.${type}`;
-  const label = strings(key);
-  if (typeof label !== 'string' || isMissingTranslation(label, key)) {
-    logMissingSportsMarketTypeTranslation(key, type);
-    return undefined;
-  }
-  return label;
-};
-
-export const getSportsMarketTypeLabel = (
-  type: string,
-  fallbackTitle?: string,
-): string =>
-  getTranslatedSportsMarketTypeLabel(type) ??
-  fallbackTitle ??
-  toTitleCase(type);
-
-type BuyHandler = (outcome: PredictOutcome, token: PredictOutcomeToken) => void;
-
-const O_U_PLAYER_PATTERN = /^(.+?):\s+\w+ O\/U/;
-
-const formatOutcomeCardTitle = (outcome: PredictOutcome): string => {
-  const raw = outcome.groupItemTitle || outcome.title;
-  if (!raw.includes('O/U')) return raw;
-
-  const match = raw.match(O_U_PLAYER_PATTERN);
-  if (match) return match[1].trim();
-
-  const colonIdx = raw.indexOf(': ');
-  if (colonIdx !== -1) return raw.slice(0, colonIdx).trim();
-
-  return raw;
-};
-
-const getTeamColor = (
-  tokenTitle: string,
-  game?: PredictMarketGame,
-): string | undefined => {
-  if (!game) return undefined;
-
-  const normalizedTokenTitle = tokenTitle.trim().toLowerCase();
-  const homeLabels = [
-    game.homeTeam.abbreviation,
-    game.homeTeam.name,
-    game.homeTeam.alias,
-  ]
-    .filter((label): label is string => Boolean(label))
-    .map((label) => label.trim().toLowerCase());
-  const awayLabels = [
-    game.awayTeam.abbreviation,
-    game.awayTeam.name,
-    game.awayTeam.alias,
-  ]
-    .filter((label): label is string => Boolean(label))
-    .map((label) => label.trim().toLowerCase());
-
-  if (homeLabels.includes(normalizedTokenTitle)) return game.homeTeam.color;
-  if (awayLabels.includes(normalizedTokenTitle)) return game.awayTeam.color;
-  return undefined;
-};
-
-const getButtonVariant = (
-  index: number,
-  total: number,
-  moneyline: boolean,
-): PredictBetButtonVariant => {
-  if (!moneyline) return 'draw';
-  if (total === 3 && index === 1) return 'draw';
-  return index === 0 ? 'yes' : 'no';
-};
-
-const buildButtons = (
-  outcome: PredictOutcome,
-  onBuyPress: BuyHandler,
-  game?: PredictMarketGame,
-  sportsMarketType?: string,
-): PredictSportOutcomeButton[] => {
-  const moneyline = isMoneylineLikeMarketType(sportsMarketType);
-  return outcome.tokens.map((token, index) => ({
-    label: token.shortTitle ?? token.title,
-    price: Math.round(token.price * 100),
-    onPress: () => onBuyPress(outcome, token),
-    variant: getButtonVariant(index, outcome.tokens.length, moneyline),
-    teamColor: moneyline
-      ? getTeamColor(token.shortTitle ?? token.title, game)
-      : undefined,
-  }));
-};
-
-const buildSubtitle = (outcome: PredictOutcome): string =>
-  `$${formatVolume(outcome.volume)} Vol`;
-
-const getDefaultLineIndex = (
-  outcomes: PredictOutcome[],
-  lineIndices: number[],
-): number => {
-  if (lineIndices.length === 0) return 0;
-
-  return lineIndices.reduce((bestIndex, outcomeIndex, lineIndex) => {
-    const bestOutcome = outcomes[lineIndices[bestIndex]];
-    const outcome = outcomes[outcomeIndex];
-    return (outcome?.volume ?? 0) > (bestOutcome?.volume ?? 0)
-      ? lineIndex
-      : bestIndex;
-  }, 0);
-};
-
-const SimpleOutcomeCard = memo(
-  ({
-    outcome,
-    title,
-    onBuyPress,
-    game,
-    sportsMarketType,
-    testID,
-  }: {
-    outcome: PredictOutcome;
-    title: string;
-    onBuyPress: BuyHandler;
-    game?: PredictMarketGame;
-    sportsMarketType?: string;
-    testID: string;
-  }) => (
-    <PredictSportOutcomeCard
-      title={title}
-      subtitle={buildSubtitle(outcome)}
-      buttons={buildButtons(outcome, onBuyPress, game, sportsMarketType)}
-      testID={testID}
-    />
-  ),
-);
-
-SimpleOutcomeCard.displayName = 'SimpleOutcomeCard';
-
-const LineOutcomeCard = memo(
-  ({
-    outcomes,
-    title,
-    onBuyPress,
-    game,
-    sportsMarketType,
-    testID,
-  }: {
-    outcomes: PredictOutcome[];
-    title?: string;
-    onBuyPress: BuyHandler;
-    game?: PredictMarketGame;
-    sportsMarketType?: string;
-    testID: string;
-  }) => {
-    const lineIndices = useMemo(
-      () =>
-        outcomes
-          .map((outcome, index) => (outcome.line != null ? index : -1))
-          .filter((index) => index !== -1),
-      [outcomes],
-    );
-
-    const lines = useMemo(
-      () => lineIndices.map((i) => Math.abs(outcomes[i].line ?? 0)),
-      [lineIndices, outcomes],
-    );
-
-    const defaultSelectedIdx = useMemo(
-      () => getDefaultLineIndex(outcomes, lineIndices),
-      [outcomes, lineIndices],
-    );
-
-    const [selectedIdx, setSelectedIdx] = useState(defaultSelectedIdx);
-
-    useEffect(() => {
-      setSelectedIdx(defaultSelectedIdx);
-    }, [defaultSelectedIdx]);
-
-    const handleSelectLine = useCallback(
-      (_line: number, indexInLines: number) => {
-        setSelectedIdx(indexInLines);
-      },
-      [],
-    );
-
-    const selectedOutcome = useMemo(
-      () => outcomes[lineIndices[selectedIdx]] ?? outcomes[0],
-      [outcomes, lineIndices, selectedIdx],
-    );
-
-    if (!selectedOutcome) return null;
-
-    return (
-      <PredictSportOutcomeCard
-        title={title ?? formatOutcomeCardTitle(selectedOutcome)}
-        subtitle={buildSubtitle(selectedOutcome)}
-        buttons={buildButtons(
-          selectedOutcome,
-          onBuyPress,
-          game,
-          sportsMarketType,
-        )}
-        lines={lines}
-        selectedLine={lines[selectedIdx]}
-        selectedIndex={selectedIdx}
-        onSelectLine={handleSelectLine}
-        testID={testID}
-      />
-    );
-  },
-);
-
-LineOutcomeCard.displayName = 'LineOutcomeCard';
-
-const isDrawOutcome = (outcome: PredictOutcome): boolean =>
-  outcome.groupItemTitle?.toLowerCase().startsWith('draw') ?? false;
-
-const sortMoneylineOutcomes = (
-  outcomes: PredictOutcome[],
-  game?: PredictMarketGame,
-): PredictOutcome[] => {
-  const hasThresholds = outcomes.some((o) => o.groupItemThreshold != null);
-  if (hasThresholds) {
-    return [...outcomes].sort(
-      (a, b) => (a.groupItemThreshold ?? 0) - (b.groupItemThreshold ?? 0),
-    );
-  }
-
-  const draw = outcomes.find(isDrawOutcome);
-  const nonDraw = outcomes.filter((o) => !isDrawOutcome(o));
-  if (!draw || nonDraw.length < 2) {
-    return [...outcomes];
-  }
-
-  if (game) {
-    const homeAbbr = game.homeTeam.abbreviation;
-    const home = nonDraw.find((o) => o.tokens[0]?.shortTitle === homeAbbr);
-    const away = nonDraw.find((o) => o !== home);
-    if (home && away) {
-      return [home, draw, away];
-    }
-  }
-
-  const sorted = [...nonDraw].sort((a, b) =>
-    (a.groupItemTitle ?? '').localeCompare(b.groupItemTitle ?? ''),
-  );
-  return [sorted[0], draw, ...sorted.slice(1)];
-};
-
-const buildMoneylineButtons = (
-  outcomes: PredictOutcome[],
-  onBuyPress: BuyHandler,
-  game?: PredictMarketGame,
-  getPrice?: ReturnType<typeof useLiveMarketPrices>['getPrice'],
-): PredictSportOutcomeButton[] => {
-  const sortedWithTokens = sortMoneylineOutcomes(outcomes, game).filter(
-    (outcome) => outcome.tokens[0] !== undefined,
-  );
-
-  return sortedWithTokens.map((outcome, i) => {
-    const yesToken = outcome.tokens[0];
-    const liveBestAsk = getPrice?.(yesToken.id)?.bestAsk;
-    const price = isValidPrice(liveBestAsk) ? liveBestAsk : yesToken.price;
-
-    return {
-      label: yesToken.shortTitle ?? yesToken.title,
-      price: Math.round(price * 100),
-      onPress: () => onBuyPress(outcome, yesToken),
-      variant: getButtonVariant(i, sortedWithTokens.length, true),
-      teamColor: getTeamColor(yesToken.shortTitle ?? yesToken.title, game),
-    };
-  });
-};
-
-const buildMoneylineSubtitle = (outcomes: PredictOutcome[]): string => {
-  const totalVolume = outcomes.reduce((sum, o) => sum + o.volume, 0);
-  return `$${formatVolume(totalVolume)} Vol`;
-};
-
-const MoneylineCard = memo(
-  ({
-    outcomes,
-    onBuyPress,
-    game,
-    title,
-    testID,
-  }: {
-    outcomes: PredictOutcome[];
-    onBuyPress: BuyHandler;
-    game?: PredictMarketGame;
-    title: string;
-    testID: string;
-  }) => {
-    const subtitle = useMemo(
-      () => buildMoneylineSubtitle(outcomes),
-      [outcomes],
-    );
-    const tokenIds = useMemo(
-      () =>
-        sortMoneylineOutcomes(outcomes, game)
-          .map((outcome) => outcome.tokens[0]?.id)
-          .filter((id): id is string => Boolean(id)),
-      [outcomes, game],
-    );
-    const { getPrice } = useLiveMarketPrices(tokenIds);
-    const buttons = useMemo(
-      () => buildMoneylineButtons(outcomes, onBuyPress, game, getPrice),
-      [outcomes, onBuyPress, game, getPrice],
-    );
-
-    return (
-      <PredictSportOutcomeCard
-        title={title}
-        subtitle={subtitle}
-        buttons={buttons}
-        buttonLayout="inlineNoSeparator"
-        testID={testID}
-      />
-    );
-  },
-);
-
-MoneylineCard.displayName = 'MoneylineCard';
-
-const SubgroupCards = memo(
-  ({
-    subgroup,
-    onBuyPress,
-    game,
-    groupKey,
-    index,
-  }: {
-    subgroup: PredictOutcomeGroup;
-    onBuyPress: BuyHandler;
-    game?: PredictMarketGame;
-    groupKey: string;
-    index: number;
-  }) => {
-    const translatedTitle = getTranslatedSportsMarketTypeLabel(subgroup.key);
-    const firstOutcomeTitle = subgroup.outcomes[0]
-      ? formatOutcomeCardTitle(subgroup.outcomes[0])
-      : undefined;
-    const title =
-      translatedTitle ?? firstOutcomeTitle ?? toTitleCase(subgroup.key);
-    const testID = `${groupKey}-${subgroup.key}-${index}`;
-
-    if (
-      isMoneylineLikeMarketType(subgroup.key) &&
-      subgroup.outcomes.length > 1
-    ) {
-      return (
-        <MoneylineCard
-          outcomes={subgroup.outcomes}
-          onBuyPress={onBuyPress}
-          game={game}
-          title={title}
-          testID={testID}
-        />
-      );
-    }
-
-    if (subgroup.outcomes.length === 1) {
-      return (
-        <SimpleOutcomeCard
-          outcome={subgroup.outcomes[0]}
-          title={title}
-          onBuyPress={onBuyPress}
-          game={game}
-          sportsMarketType={subgroup.key}
-          testID={testID}
-        />
-      );
-    }
-
-    return (
-      <LineOutcomeCard
-        outcomes={subgroup.outcomes}
-        title={translatedTitle}
-        onBuyPress={onBuyPress}
-        game={game}
-        sportsMarketType={subgroup.key}
-        testID={testID}
-      />
-    );
-  },
-);
-
-SubgroupCards.displayName = 'SubgroupCards';
-
-const OutcomesContent = memo(
-  ({
-    group,
-    onBuyPress,
-    game,
-  }: {
-    group: PredictOutcomeGroup;
-    onBuyPress: BuyHandler;
-    game?: PredictMarketGame;
-  }) => {
-    if (group.subgroups && group.subgroups.length > 0) {
-      return (
-        <>
-          {group.subgroups.map((subgroup, index) => (
-            <SubgroupCards
-              key={subgroup.key}
-              subgroup={subgroup}
-              onBuyPress={onBuyPress}
-              game={game}
-              groupKey={group.key}
-              index={index}
-            />
-          ))}
-        </>
-      );
-    }
-
-    const firstType = group.outcomes[0]?.sportsMarketType;
-    if (
-      firstType &&
-      isMoneylineLikeMarketType(firstType) &&
-      group.outcomes.length > 1
-    ) {
-      return (
-        <MoneylineCard
-          outcomes={group.outcomes}
-          onBuyPress={onBuyPress}
-          game={game}
-          title={getSportsMarketTypeLabel(
-            firstType,
-            formatOutcomeCardTitle(group.outcomes[0]),
-          )}
-          testID={`${group.key}-moneyline`}
-        />
-      );
-    }
-
-    return (
-      <>
-        {group.outcomes.map((outcome, index) => (
-          <SimpleOutcomeCard
-            key={outcome.id}
-            outcome={outcome}
-            title={formatOutcomeCardTitle(outcome)}
-            onBuyPress={onBuyPress}
-            game={game}
-            sportsMarketType={outcome.sportsMarketType}
-            testID={`${group.key}-outcome-${index}`}
-          />
-        ))}
-      </>
-    );
-  },
-);
-
-OutcomesContent.displayName = 'OutcomesContent';
+export { getSportsMarketTypeLabel } from './utils';
 
 export interface OutcomesTabProps {
   groupMap: Map<string, PredictOutcomeGroup>;
@@ -605,49 +21,8 @@ export interface OutcomesTabProps {
 
 const PredictGameOutcomesTab = memo(
   ({ groupMap, game, activeChipKey, onBuyPress }: OutcomesTabProps) => {
-    const { isBuySheetOpen } = usePredictPreviewSheet();
     const selectedGroup = groupMap.get(activeChipKey);
-
-    const flatOutcomes = useMemo(
-      () => (selectedGroup ? flattenGroupOutcomes(selectedGroup) : []),
-      [selectedGroup],
-    );
-
-    const priceQueries = useMemo(
-      () => buildPriceQueries(flatOutcomes),
-      [flatOutcomes],
-    );
-
-    const shouldFetchPrices = priceQueries.length > 0 && !isBuySheetOpen;
-    const activePriceQueries = shouldFetchPrices
-      ? priceQueries
-      : EMPTY_PRICE_QUERIES;
-
-    const { prices } = usePredictPrices({
-      queries: activePriceQueries,
-      enabled: shouldFetchPrices,
-      pollingInterval: shouldFetchPrices ? PRICE_POLL_INTERVAL_MS : undefined,
-    });
-
-    const previousSelectedGroupRef = useRef<PredictOutcomeGroup | null>(null);
-    const previousPricedOutcomesRef = useRef<PredictOutcome[]>([]);
-
-    const pricedGroup = useMemo(() => {
-      if (!selectedGroup) {
-        return undefined;
-      }
-
-      const nextGroup = applyPricesToGroup(
-        selectedGroup,
-        prices,
-        previousSelectedGroupRef.current,
-        previousPricedOutcomesRef.current,
-      );
-
-      previousSelectedGroupRef.current = selectedGroup;
-      previousPricedOutcomesRef.current = flattenGroupOutcomes(nextGroup);
-      return nextGroup;
-    }, [selectedGroup, prices]);
+    const pricedGroup = usePricedOutcomeGroup(selectedGroup);
 
     return (
       <Box testID={PREDICT_GAME_DETAILS_CONTENT_TEST_IDS.OUTCOMES_CONTENT}>
