+++ app/components/UI/Predict/utils/resolvePredictFeatureFlags.ts
@@ -10,7 +10,7 @@ import {
   DEFAULT_PREDICT_WORLD_CUP_FLAG,
 } from '../constants/flags';
 import { filterSupportedLeagues } from '../constants/sports';
-import { SUPPORTED_SPORTS_MARKET_TYPES } from '../providers/polymarket/constants';
+import { normalizeEnabledSportsMarketTypes } from '../providers/polymarket/outcomeGrouping';
 import {
   parse,
   PredictFeeCollectionSchema,
@@ -39,14 +39,6 @@ function resolveVersionGatedBooleanFlag(
     ) ?? fallback
   );
 }
-
-const filterSupportedSportsMarketTypes = (types: string[] = []): string[] => [
-  ...new Set(
-    types
-      .map((type) => type.toLowerCase())
-      .filter((type) => SUPPORTED_SPORTS_MARKET_TYPES.has(type)),
-  ),
-];
 
 /**
  * Resolves the Predict feature flags used by both the controller and selectors.
@@ -100,10 +92,8 @@ export function resolvePredictFeatureFlags(
     ? filterSupportedLeagues(extendedSportsFlag.leagues ?? [])
     : [];
   const enabledSportsMarketTypes = extendedSportsMarketsEnabled
-    ? filterSupportedSportsMarketTypes(
-        Array.isArray(extendedSportsFlag.enabledSportsMarketTypes)
-          ? extendedSportsFlag.enabledSportsMarketTypes
-          : [],
+    ? normalizeEnabledSportsMarketTypes(
+        extendedSportsFlag.enabledSportsMarketTypes,
       )
     : [];
   const fakOrdersEnabled = resolveVersionGatedBooleanFlag(
